fix: remove unused rich dep and fix timing precision in records
- Remove unused runtime dependency 'rich' from pyproject.toml and regenerate uv.lock
- Switch timing fields from int ms to float ms in Stats, RecordEntry, engine, and session
- Fixes records detail page showing 0 for total/average turn time when single turns take <1ms
- Update tests to match float timing values
- Add AGENTS.md project guide

+++ src/microciv/game/engine.py
@@ -70,7 +70,7 @@ class GameEngine:
         else:
             self.state.turn += 1
 
-        elapsed_ms = int((perf_counter() - started_at) * 1000)
+        elapsed_ms = (perf_counter() - started_at) * 1000
         self.state.stats.record_turn_time(elapsed_ms)
         return EngineResult(True, "", self.state, settlement)
